pyproject.toml
[project]
name = "langfuse-agent-poc"
version = "0.1.0"
description = "Langfuse Agent POC"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "langfuse",
    "google-genai",
    "langfuse",
    "python-dotenv",
    "langgraph",
    "langchain",
    "langchain-google-genai"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

